MCP-1623: moved and re-enabled architecture tests
